Update based on style guide
